+++ packages/migrations/src/Command/MigrateCommand.php
@@ -2,18 +2,13 @@
 
 namespace Shopsys\MigrationBundle\Command;
 
-use Doctrine\DBAL\Migrations\Configuration\Configuration;
-use Doctrine\DBAL\Migrations\OutputWriter;
-use Doctrine\DBAL\Migrations\Tools\Console\Helper\ConfigurationHelper;
 use Doctrine\ORM\EntityManagerInterface;
-use Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator;
-use Symfony\Component\Console\Command\Command;
+use Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLock;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
-class MigrateCommand extends Command
+class MigrateCommand extends AbstractCommand
 {
     const RETURN_CODE_OK = 0;
     const RETURN_CODE_ERROR = 1;
@@ -29,28 +24,16 @@ class MigrateCommand extends Command
     private $em;
 
     /**
-     * @var \Symfony\Component\DependencyInjection\ContainerInterface
+     * @var \Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLock
      */
-    private $container;
+    private $migrationsLock;
 
-    /**
-     * @var \Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator
-     */
-    private $migrationsLocator;
-
-    /**
-     * @param \Doctrine\ORM\EntityManagerInterface $em
-     * @param \Symfony\Component\DependencyInjection\ContainerInterface $container
-     * @param \Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationsLocator $migrationsLocator
-     */
     public function __construct(
         EntityManagerInterface $em,
-        ContainerInterface $container,
-        MigrationsLocator $migrationsLocator
+        MigrationsLock $migrationsLock
     ) {
         $this->em = $em;
-        $this->container = $container;
-        $this->migrationsLocator = $migrationsLocator;
+        $this->migrationsLock = $migrationsLock;
 
         parent::__construct();
     }
@@ -70,48 +53,20 @@ class MigrateCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         try {
-            $migrationsConfiguration = $this->createAndRegisterMigrationsConfiguration($output);
+            $this->em->transactional(function () use ($output) {
+                $this->executeDoctrineMigrateCommand($output);
 
-            foreach ($this->migrationsLocator->getMigrationsLocations() as $migrationsLocation) {
-                $migrationsConfiguration->setMigrationsDirectory($migrationsLocation->getDirectory());
-                $migrationsConfiguration->setMigrationsNamespace($migrationsLocation->getNamespace());
+                $output->writeln('');
 
-                $output->writeln('Installing migrations from ' . $migrationsLocation->getDirectory() . ' in namespace ' . $migrationsLocation->getNamespace());
-
-                $this->em->transactional(function () use ($output) {
-                    $this->executeDoctrineMigrateCommand($output);
-
-                    $output->writeln('');
-                });
-            }
-
-            $output->writeln('Migrations from all sources has been installed.');
-
-            $this->executeCheckSchemaCommand($output);
+                $this->executeCheckSchemaCommand($output);
+            });
         } catch (\Exception $ex) {
-            $message = "Database migration process did not run properly. Last transaction was reverted.\n"
-                . 'For more informations see the previous exception.';
+            $message = 'Database migration process did not run properly. Transaction was reverted.';
             throw new \Shopsys\MigrationBundle\Command\Exception\MigrateCommandException($message, $ex);
         }
-    }
 
-    /**
-     * @param \Symfony\Component\Console\Output\OutputInterface $output
-     * @return \Doctrine\DBAL\Migrations\Configuration\Configuration
-     */
-    private function createAndRegisterMigrationsConfiguration(OutputInterface $output)
-    {
-        $outputWriter = new OutputWriter(
-            function ($message) use ($output) {
-                return $output->writeln($message);
-            }
-        );
-
-        $migrationsConfiguration = new Configuration($this->em->getConnection(), $outputWriter);
-        $configurationHelper = new ConfigurationHelper($this->em->getConnection(), $migrationsConfiguration);
-        $this->getApplication()->getHelperSet()->set($configurationHelper, 'configuration');
-
-        return $migrationsConfiguration;
+        $migrationVersions = $this->getMigrationsConfiguration()->getMigrations();
+        $this->migrationsLock->saveNewMigrations($migrationVersions);
     }
 
     /**